fix: guard docker gid detection by platform
